Functional Tests for group PHP71 have been dropped
