+++ src/Test/Functional/Acceptance/Cron22Cest.php
@@ -8,7 +8,7 @@ declare(strict_types=1);
 namespace Magento\MagentoCloud\Test\Functional\Acceptance;
 
 /**
- * @group php71
+ * @group php72
  */
 class Cron22Cest extends CronCest
 {
@@ -19,7 +19,7 @@ class Cron22Cest extends CronCest
     {
         return [
             [
-                'version' => '2.2.0',
+                'version' => '2.2.10',
                 'variables' => [
                     'MAGENTO_CLOUD_VARIABLES' => [
                         'ADMIN_EMAIL' => 'admin@example.com',
@@ -28,32 +28,14 @@ class Cron22Cest extends CronCest
                 ],
             ],
             [
-                'version' => '2.2.4',
+                'version' => '2.2.11',
                 'variables' => [
                     'MAGENTO_CLOUD_VARIABLES' => [
                         'ADMIN_EMAIL' => 'admin@example.com',
                         'ADMIN_LOCALE' => 'ar_KW'
                     ],
                 ],
-            ],
-            [
-                'version' => '2.2.5',
-                'variables' => [
-                    'MAGENTO_CLOUD_VARIABLES' => [
-                        'ADMIN_EMAIL' => 'admin@example.com',
-                        'ADMIN_LOCALE' => 'ar_KW'
-                    ],
-                ],
-            ],
-            [
-                'version' => '2.2.8',
-                'variables' => [
-                    'MAGENTO_CLOUD_VARIABLES' => [
-                        'ADMIN_EMAIL' => 'admin@example.com',
-                        'ADMIN_LOCALE' => 'fr_FR'
-                    ],
-                ],
-            ],
+            ]
         ];
     }
 }
