+++ tests/functional/_data/modules/Magento/CronTest/composer.json
@@ -2,7 +2,7 @@
     "name": "magento/module-cron-test",
     "description": "N/A",
     "require": {
-        "php": "7.0.2|7.0.4|~7.0.6|~7.1.0"
+        "php": "~7.2.0"
     },
     "type": "magento2-module",
     "version": "100.2.0",
